Move contact selectors to the proper directory
[#22]
