+++ app/javascript/selectors/investigation/contactLogSelectors.js
